requirements.txt
Jinja2==2.7.3
ExifRead==1.4.2
numpy==1.8.1
Pillow==2.5.1
